Refactored query responsibility from clerk and ictoffcier to QueryRole classes Refactored database functionality interfaces to IUpdate or IQuery Created QueryRole classes that handles the application level usage of dao classes
