--- src/main/java/management/IRegistering.java
@@ -1,8 +1,0 @@
-package management;
-
-public interface IRegistering<T extends Person> {
-    boolean registerPerson(T person, Department department);
-
-    boolean isPersonRegistered(T person, Department department);
-
-}
