--- src/main/java/storage/IDelete.java
@@ -1,7 +1,0 @@
-package storage;
-
-import management.Department;
-
-public interface IDelete<T> {
-    boolean delete(T obj, Department department);
-}
