+++ src/main/java/storage/IQueryPatients.java
@@ -1,0 +1,7 @@
+package storage;
+
+import management.Patient;
+
+public interface IQueryPatients<T extends Patient> extends IQuery<T> {
+}
+
