+++ src/main/java/storage/IQueryStaff.java
@@ -1,0 +1,6 @@
+package storage;
+
+import management.Staff;
+
+public interface IQueryStaff<T extends Staff> extends IQuery<T> {
+}
