+++ src/main/java/storage/IRegistering.java
@@ -1,0 +1,5 @@
+package storage;
+
+public interface IRegistering {
+
+}
