+++ src/main/java/storage/IUpdate.java
@@ -1,0 +1,12 @@
+package storage;
+
+import management.Department;
+import management.Person;
+
+public interface IUpdate {
+    <T extends Person> boolean registerPerson(T person, Department department);
+
+    <T extends Person> boolean isPersonRegistered(T person, Department department);
+
+    <T> boolean delete(T obj, Department department);
+}
